fix: リソースリーク・NULL 参照・UB を一括修正
コードレビューで検出された複数のバグを一括修正。
メモリリーク（ワーカースレッドキュー、スレッド配列、デコーダバッファ）、
GDI DC リーク（AnimateTransition、PanelRect）、NULL 参照保護漏れ、
画像サイズ計算の int オーバーフロー（各デコーダを size_t 化）、
C++ UB（非 virtual デストラクタ、TGA 自己代入、
EXIFReader 未定義戻り値）、イテレータ無効化、
INVALID_HANDLE_VALUE 比較ミス、
JXL キャッシュの new[]/free ミスマッチを修正。

+++ src/JPEGView/PNGWrapper.cpp
@@ -152,11 +152,11 @@ void* PngReader::ReadNextFrame(void** exif_chunk, png_uint_32* exif_size)
 		for (j = 0; j < cache.h0; j++)
 			memcpy(cache.rows_image[j + cache.y0] + cache.x0 * 4, cache.rows_frame[j], cache.w0 * 4);
 
-	void* pixels = malloc(cache.width * cache.height * cache.channels);
+	void* pixels = malloc((size_t)cache.width * cache.height * cache.channels);
 	if (pixels == NULL)
 		return NULL;
 	for (j = 0; j < cache.height; j++)
-		memcpy((char*)pixels + j * cache.width * cache.channels, cache.rows_image[j], cache.width * cache.channels);
+		memcpy((char*)pixels + (size_t)j * cache.width * cache.channels, cache.rows_image[j], (size_t)cache.width * cache.channels);
 
 #ifdef PNG_APNG_SUPPORTED
 	if (cache.dop == PNG_DISPOSE_OP_PREVIOUS)
